+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -226,7 +226,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K475ZJ11Z3VB26R76NCJMR32"
+        "01K47B8VC86T1V1DJZC2B8P1X0"
     )
     r"""The id of the resource"""